+++ acumos/tests/test_session.py
@@ -23,6 +23,7 @@ import contextlib
 import mock
 import tempfile
 import json
+import sys
 from os import environ, listdir
 from os.path import isfile, join as path_join
 
@@ -37,18 +38,21 @@ from jsonschema import validate
 
 from acumos.wrapped import _infer_model_dir
 from acumos.modeling import Model, List, create_namedtuple, create_dataframe
-from acumos.session import AcumosSession, _dump_model, Requirements
+from acumos.session import AcumosSession, _dump_model
 from acumos.exc import AcumosError
 from acumos.utils import load_artifact
 from acumos.auth import clear_jwt, _USERNAME_VAR, _PASSWORD_VAR, _TOKEN_VAR
-from acumos.metadata import SCHEMA_VERSION
+from acumos.metadata import SCHEMA_VERSION, Requirements
 
 from mock_server import MockServer
-from utils import TEST_DIR
+from utils import run_command, TEST_DIR
+from user_package import user_package_module
+from user_module import user_function
 
 
+_USER_PACKAGE_DIR = path_join(TEST_DIR, 'user_package')
+_MODEL_LOADER_HELPER = path_join(TEST_DIR, 'model_loader_helper.py')
 _REQ_FILES = ('model.zip', 'model.proto', 'metadata.json')
-_CUSTOM_PACKAGE_DIR = path_join(TEST_DIR, 'custom_package')
 _FAKE_USERNAME = 'foo'
 _FAKE_PASSWORD = 'bar'
 _FAKE_TOKEN = 'secrettoken'
@@ -81,25 +85,71 @@ def test_extra_header():
 
 
 def test_custom_package():
-    '''Tests that custom packages can be included'''
-    from custom_package import custom_module  # local to tests/ directory
+    '''Tests that custom packages can be included, wrapped, and loaded'''
 
     def my_transform(x: int, y: int) -> int:
-        return custom_module.add_numbers(x, y)
+        return user_package_module.add_numbers(x, y)
 
     model = Model(transform=my_transform)
     model_name = 'my-model'
 
-    reqs = Requirements(packages=[_CUSTOM_PACKAGE_DIR])
+    # load should fail without requirements
+    with pytest.raises(Exception, match='Module user_package was detected as a dependency'):
+        with _dump_model(model, model_name) as dump_dir:
+            pass
 
-    s = AcumosSession()
+    reqs = Requirements(packages=[_USER_PACKAGE_DIR])
 
-    with tempfile.TemporaryDirectory() as tdir:
+    with _dump_model(model, model_name, reqs) as dump_dir:
+        run_command([sys.executable, _MODEL_LOADER_HELPER, dump_dir, 'user_package'])
 
-        with pytest.raises(AcumosError):
-            s.dump(model, model_name, tdir)  # fails because `custom_package` is not declared
 
-        s.dump(model, model_name, tdir, reqs)
+def test_custom_script():
+    '''Tests that custom modules can be included, wrapped, and loaded'''
+
+    def predict(x: int) -> int:
+        return user_function(x)
+
+    model = Model(predict=predict)
+    model_name = 'my-model'
+
+    with _dump_model(model, model_name) as dump_dir:
+        run_command([sys.executable, _MODEL_LOADER_HELPER, dump_dir, 'user_module'])
+
+
+def test_script_req():
+    '''Tests that Python scripts can be included using Requirements'''
+
+    def predict(x: int) -> int:
+        return x
+
+    model = Model(predict=predict)
+    model_name = 'my-model'
+
+    # tests that individual script and directory of scripts are both gathered
+    reqs = Requirements(scripts=_abspath('./user_module.py', './user_package'))
+
+    with _dump_model(model, model_name, reqs) as dump_dir:
+        _verify_files(dump_dir, ('scripts/user_provided/user_package_module.py',
+                                 'scripts/user_provided/__init__.py',
+                                 'scripts/user_provided/user_module.py'))
+
+    bad_reqs = Requirements(scripts=_abspath('./user_module.py', './user_package', 'not_real.py'))
+
+    with pytest.raises(AcumosError, match='does not exist'):
+        with _dump_model(model, model_name, bad_reqs) as dump_dir:
+            pass
+
+    bad_reqs = Requirements(scripts=_abspath('./user_module.py', './user_package', './att.png'))
+
+    with pytest.raises(AcumosError, match='is invalid'):
+        with _dump_model(model, model_name, bad_reqs) as dump_dir:
+            pass
+
+
+def _abspath(*files):
+    '''Returns the absolute path of a local test files'''
+    return tuple(path_join(TEST_DIR, file) for file in files)
 
 
 def test_session_dump():
@@ -126,13 +176,13 @@ def test_session_dump():
 def test_dump_model():
     '''Tests dump model utility, including generated artifacts'''
 
-    def func() -> None:
-        pass
+    def predict(x: int) -> int:
+        return user_function(x)
 
-    model = Model(transform=func)
+    model = Model(predict=predict)
     model_name = 'my-model'
 
-    reqs = Requirements(reqs=['wronglib'], req_map={'wronglib': 'scipy'}, packages=[_CUSTOM_PACKAGE_DIR])
+    reqs = Requirements(reqs=['wronglib'], req_map={'wronglib': 'scipy'}, packages=[_USER_PACKAGE_DIR])
 
     with _dump_model(model, model_name, reqs) as dump_dir:
 
@@ -145,10 +195,17 @@ def test_dump_model():
         # test that a user-provided library was included and correctly mapped
         assert 'scipy' in {r['name'] for r in metadata['runtime']['dependencies']['pip']['requirements']}
 
-        # test that custom code was bundled
-        model_dir = _infer_model_dir(dump_dir)
-        assert isfile(path_join(model_dir, 'scripts', 'user_provided', 'custom_package', 'custom_module.py'))
-        assert isfile(path_join(model_dir, 'scripts', 'user_provided', 'custom_package', '__init__.py'))
+        # test that custom package was bundled
+        _verify_files(dump_dir, ('scripts/user_provided/user_package/user_package_module.py',
+                                 'scripts/user_provided/user_package/__init__.py',
+                                 'scripts/user_provided/user_module.py'))
+
+
+def _verify_files(dump_dir, files):
+    '''Asserts that `files` exist in `dump_dir`'''
+    model_dir = _infer_model_dir(dump_dir)
+    for file in files:
+        assert isfile(path_join(model_dir, file))
 
 
 def _load_schema(version):
@@ -187,8 +244,8 @@ def test_session_push_sklearn():
 
             model = Model(predict=predict)
 
-            model_uri, auth_uri, _, _ = server.config
-            s = AcumosSession(model_uri, auth_uri)
+            model_url, auth_url, _, _ = server.config
+            s = AcumosSession(model_url, auth_url)
             s.push(model, name='sklearn_iris_push')
 
 
@@ -223,8 +280,8 @@ def test_session_push_keras():
 
             model = Model(predict=predict)
 
-            model_uri, auth_uri, _, _ = server.config
-            s = AcumosSession(model_uri, auth_uri)
+            model_url, auth_url, _, _ = server.config
+            s = AcumosSession(model_url, auth_url)
             s.push(model, name='keras_iris_push')
 
 
@@ -237,8 +294,8 @@ def _push_dummy_model(extra_headers=None):
     model = Model(transform=my_transform)
 
     with MockServer() as server:
-        model_uri, auth_uri, _, _ = server.config
-        s = AcumosSession(model_uri, auth_uri)
+        model_url, auth_url, _, _ = server.config
+        s = AcumosSession(model_url, auth_url)
         s.push(model, name='my-model', extra_headers=extra_headers)